Spit MAD at TestCollisions
